+++ packages/fiori/src/SideNavigationItem.ts
@@ -26,6 +26,7 @@ import type SideNavigationSubItem from "./SideNavigationSubItem.js";
  * @abstract
  * @tagname ui5-side-navigation-item
  * @public
+ * @implements sap.ui.webc.fiori.ISideNavigationItem
  * @since 1.0.0-rc.8
  */
 @customElement("ui5-side-navigation-item")
@@ -55,7 +56,7 @@ class SideNavigationItem extends UI5Element {
 	 *
 	 * The SAP-icons font provides numerous options.
 	 * <br>
-	 * See all the available icons in the <ui5-link target="_blank" href="https://sdk.openui5.org/test-resources/sap/m/demokit/iconExplorer/webapp/index.html" class="api-table-content-cell-link">Icon Explorer</ui5-link>.
+	 * See all the available icons in the <ui5-link target="_blank" href="https://sdk.openui5.org/test-resources/sap/m/demokit/iconExplorer/webapp/index.html">Icon Explorer</ui5-link>.
 	 * @public
 	 * @type {string}
 	 * @defaultvalue ""