+++ packages/fiori/src/types/MediaGalleryMenuHorizontalAlign.ts
@@ -1,4 +1,6 @@
 /**
+ * Defines the horizontal alignment of the thumbnails menu of the <code>ui5-media-gallery</code> component.
+ *
  * @readonly
  * @enum {string}
  * @public