+++ packages/fiori/test/specs/ProductSwitchItem.spec.js
@@ -13,3 +13,12 @@ describe("ProductSwitchItem general interaction", async () => {
 		assert.ok(await productSwitchItem.shadow$(".ui5-product-switch-item-subtitle"), "SubTitle is rendered.");
 	});
 });
+
+describe("ARIA attributes", () => {
+	it("items role set correctly", async () => {
+		const productSwitchItem = await browser.$("#productSwitchItem");
+		const root = await productSwitchItem.shadow$(".ui5-product-switch-item-root");
+		const role = await root.getAttribute("role");
+		assert.strictEqual(role, "listitem", "should have role listitem");
+	});
+});