+++ packages/main/src/MenuItem.ts
@@ -62,7 +62,7 @@ class MenuItem extends UI5Element {
 	 * The SAP-icons font provides numerous options.
 	 * <br><br>
 	 <b>* Example:</b>
-	 * See all the available icons in the <ui5-link target="_blank" href="https://sdk.openui5.org/test-resources/sap/m/demokit/iconExplorer/webapp/index.html" class="api-table-content-cell-link">Icon Explorer</ui5-link>.
+	 * See all the available icons in the <ui5-link target="_blank" href="https://sdk.openui5.org/test-resources/sap/m/demokit/iconExplorer/webapp/index.html">Icon Explorer</ui5-link>.
 	 *
 	 * @name sap.ui.webc.main.MenuItem.prototype.icon
 	 * @type {string}