+++ packages/main/test/specs/DatePicker.spec.js
@@ -1133,7 +1133,21 @@ describe("Date Picker Tests", () => {
 		await root.keys("Jan 1, 1999999");
 		await browser.$("#dp5").shadow$("ui5-input").shadow$("input").click(); //click elsewhere to focusout
 
+		assert.equal(await input.getProperty("value"), "", 'the value is not changed');
+	});
+
+	it("when value is prevented, the target value is restored to previous one", async () => {
+		datepicker.id = "#dpPrevent";
+
+		const input = await datepicker.getInput();
+		await input.click();
+
+		const root = await datepicker.getRoot();
+		await root.keys("Mar 31, 1995");
+		await browser.$("#dp5").shadow$("ui5-input").shadow$("input").click(); //click elsewhere to focusout
+
 		assert.equal(await input.getProperty("valueState"), "None", 'the value state is not changed');
+		assert.equal(await browser.$("#lblChangePreventTargetValue").getHTML(false), "target value on event fire: Mar 31, 1995", 'the value is applied on target before prevention');
 	});
 
 	it("DatePicker's formatter has strict parsing enabled", async () => {
@@ -1192,4 +1206,86 @@ describe("Date Picker Tests", () => {
 
 		assert.equal(await input.getProperty("valueState"), "Error", "value state of the input is valid (3)");
 	});
+
+	it("should open calendar picker in CalendarMode.DAY_MONTH_YEAR mode", async () => {
+		datepicker.id = "#dpCalendarModeDays";
+		const innerInput = await datepicker.getInnerInput();
+		await innerInput.click();
+		await browser.keys("2020, 04/01");
+		await browser.keys("Enter");
+
+		await datepicker.openPicker();
+		const calendar = await datepicker.getCalendar();
+
+		const currentPicker = await calendar.getProperty("_currentPicker");
+		assert.equal(currentPicker, "day", "calendar is opened on days");
+
+		const dayPicker = await calendar.shadow$("ui5-daypicker");
+		const monthPicker = await calendar.shadow$("ui5-monthpicker");
+		const yearPicker = await calendar.shadow$("ui5-yearpicker");
+		assert.notOk(await dayPicker.getAttribute("hidden"));
+		assert.ok(await monthPicker.getAttribute("hidden"));
+		assert.ok(await yearPicker.getAttribute("hidden"));
+
+		const timestamp_30_Jan_2020 = 1580342400;
+		const calendarDate_30_Jan_2020 = await datepicker.getPickerDate(timestamp_30_Jan_2020);
+		await calendarDate_30_Jan_2020.click();
+
+		assert.isFalse(await datepicker.isPickerOpen(), "picker is closed after day selection");
+	});
+
+	it("should open calendar picker in CalendarMode.MONTH_YEAR mode", async () => {
+		datepicker.id = "#dpCalendarModeMonths";
+		const innerInput = await datepicker.getInnerInput();
+		await innerInput.click();
+		await browser.keys("June 2023");
+		await browser.keys("Enter");
+
+		const calendar = await datepicker.getCalendar();
+		await datepicker.openPicker();
+
+		const currentPicker = await calendar.getProperty("_currentPicker");
+		assert.equal(currentPicker, "month", "calendar is opened on months");
+
+		const dayPicker = await calendar.shadow$("ui5-daypicker");
+		const monthPicker = await calendar.shadow$("ui5-monthpicker");
+		const yearPicker = await calendar.shadow$("ui5-yearpicker");
+		assert.ok(await dayPicker.getAttribute("hidden"));
+		assert.notOk(await monthPicker.getAttribute("hidden"));
+		assert.ok(await yearPicker.getAttribute("hidden"));
+
+		const timestamp_Nov_2023 = 1698796800;
+		const calendarDate_Nov_2023 = await datepicker.getPickerMonth(timestamp_Nov_2023);
+		await calendarDate_Nov_2023.click();
+
+		assert.isFalse(await datepicker.isPickerOpen(), "picker is closed after month selection");
+	});
+
+	it("should open calendar picker in CalendarMode.YEAR mode", async () => {
+		datepicker.id = "#dpCalendarModeYears";
+		const innerInput = await datepicker.getInnerInput();
+		await innerInput.click();
+		await browser.keys("2018");
+		await browser.keys("Enter");
+
+		await datepicker.openPicker();
+		const calendar = await datepicker.getCalendar();
+
+		const currentPicker = await calendar.getProperty("_currentPicker");
+		assert.equal(currentPicker, "year", "calendar is opened on months");
+
+		const dayPicker = await calendar.shadow$("ui5-daypicker");
+		const monthPicker = await calendar.shadow$("ui5-monthpicker");
+		const yearPicker = await calendar.shadow$("ui5-yearpicker");
+		assert.ok(await dayPicker.getAttribute("hidden"));
+		assert.ok(await monthPicker.getAttribute("hidden"));
+		assert.notOk(await yearPicker.getAttribute("hidden"));
+
+		const timestamp_2014 = 1388534400;
+		const calendarDate_2014 = await datepicker.getPickerYear(timestamp_2014);
+		await calendarDate_2014.click();
+
+		assert.isFalse(await datepicker.isPickerOpen(), "picker is closed after year selection");
+	});
+
 });