+++ packages/main/test/specs/Popover.spec.js
@@ -406,3 +406,109 @@ describe("Acc", () => {
 		assert.isNull(await popoverWithoutHeader.shadow$(".ui5-popup-root").getAttribute("aria-labelledby"), "Popover should NOT have aria-labelledby set.");
 	});
 });
+
+describe("Horizontal Alignment", () => {
+	before(async () => {
+		await browser.url(`test/pages/Popover.html`);
+	});
+
+	const EPS = 2; // 2px
+
+	const isHorizontallyCentered = async (popover, opener) => {
+		const popoverRect = {
+			...await popover.getLocation(),
+			...await popover.getSize()
+		};
+		const openerRect = {
+			...await opener.getLocation(),
+			...await opener.getSize()
+		};
+		const openerCenter = openerRect.x + openerRect.width / 2;
+		const expectedPopoverLeft = openerCenter - popoverRect.width / 2;
+
+		return Math.abs(popoverRect.x - expectedPopoverLeft) < EPS;
+	}
+
+	const isHorizontallyLeftAligned = async (popover, opener) => {
+		const popoverRect = {
+			...await popover.getLocation(),
+			...await popover.getSize()
+		};
+		const openerRect = {
+			...await opener.getLocation(),
+			...await opener.getSize()
+		};
+
+		return Math.abs(openerRect.x - popoverRect.x) < EPS;
+	}
+
+	const isHorizontallyRightAligned = async (popover, opener) => {
+		const popoverRect = {
+			...await popover.getLocation(),
+			...await popover.getSize()
+		};
+		const openerRect = {
+			...await opener.getLocation(),
+			...await opener.getSize()
+		};
+		const openerRight = openerRect.x + openerRect.width;
+		const popoverRight = popoverRect.x + popoverRect.width;
+
+		return Math.abs(openerRight - popoverRight) < EPS;
+	}
+
+	it("Center", async () => {
+		await browser.$("[ui5-radio-button][name='horizontalAlign'][text='Center']").click();
+		await browser.$("#horizontalAlignBtn").click();
+		const popover = await browser.$("#popoverHorizontalAlign");
+		const opener = await browser.$("#targetOpener");
+	
+		assert.ok(await isHorizontallyCentered(popover, opener), `Popover should be centered`);
+	});
+
+	it("Left", async () => {
+		await browser.$("[ui5-radio-button][name='horizontalAlign'][text='Left']").click();
+		await browser.$("#horizontalAlignBtn").click();
+		const popover = await browser.$("#popoverHorizontalAlign");
+		const opener = await browser.$("#targetOpener");
+
+		assert.ok(await isHorizontallyLeftAligned(popover, opener), `Popover should be left aligned`);
+	});
+
+	it("Right", async () => {
+		await browser.$("[ui5-radio-button][name='horizontalAlign'][text='Right']").click();
+		await browser.$("#horizontalAlignBtn").click();
+		const popover = await browser.$("#popoverHorizontalAlign");
+		const opener = await browser.$("#targetOpener");
+	
+		assert.ok(await isHorizontallyRightAligned(popover, opener), `Popover should be right aligned`);
+	});
+
+	it("Center, in RTL", async () => {
+		await browser.$("[ui5-radio-button][name='horizontalAlign'][text='Center']").click();
+		await browser.$("#rtlCb").click();
+		await browser.$("#horizontalAlignBtn").click();
+		const popover = await browser.$("#popoverHorizontalAlign");
+		const opener = await browser.$("#targetOpener");
+
+		assert.ok(await isHorizontallyCentered(popover, opener), `Popover should be centered`);
+	});
+
+	it("Left, in RTL", async () => {
+		await browser.$("[ui5-radio-button][name='horizontalAlign'][text='Left']").click();
+		await browser.$("#horizontalAlignBtn").click();
+		const popover = await browser.$("#popoverHorizontalAlign");
+		const opener = await browser.$("#targetOpener");
+
+		assert.ok(isHorizontallyRightAligned(popover, opener), `Popover should be right aligned, flipped by RTL direction`);
+	});
+
+	it("Right, in RTL", async () => {
+		await browser.$("[ui5-radio-button][name='horizontalAlign'][text='Right']").click();
+		await browser.$("#horizontalAlignBtn").click();
+		const popover = await browser.$("#popoverHorizontalAlign");
+		const opener = await browser.$("#targetOpener");
+
+		assert.ok(await isHorizontallyLeftAligned(popover, opener), `Popover should be left aligned, flipped by RTL direction`);
+	});
+});